refactor(multitenancy): move Tenant + definitions to base module (ADR-020)
Tenant entity, TenantQueryDefinition and TenantExportDefinition relocate
to Granit.MultiTenancy. Allows host applications to use the framework
TenantExportDefinition (replacing local duplicates) and aligns the
module with the ADR-020 placement convention.
